fix(support-plan-guide): resolve chunk size limit and update capability test
